requirements.txt
mediapipe~=0.10.9
pillow>=10.2.0
smbus2~=0.4.3
spidev==3.6
numpy>=1.26.2